pyproject.toml
[build-system]
requires = ["poetry-core>=2"]
build-backend = "poetry.core.masonry.api"

[project]
name = "qrcode"
version = "8.2"
description = "QR Code image generator"
authors = [
  { name = "Lincoln Loop", email = "info@lincolnloop.com" },
]
license = { text = "BSD-3-Clause" }
dynamic = ["readme"]
keywords = ["qr", "denso-wave", "IEC18004"]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Multimedia :: Graphics",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = "~=3.9"
dependencies = [
  "colorama; sys_platform == 'win32'",
  "deprecation",
]


[project.optional-dependencies]
pil = ["pillow >=9.1.0"]
png = ["pypng"]
all = ["pypng", "pillow >=9.1.0"]

[project.urls]
homepage = "https://github.com/lincolnloop/python-qrcode"
repository = "https://github.com/lincolnloop/python-qrcode.git"
documentation = "https://github.com/lincolnloop/python-qrcode#readme"
changelog = "https://github.com/lincolnloop/python-qrcode/blob/main/CHANGES.rst"
"Bug Tracker" = "https://github.com/lincolnloop/python-qrcode/issues"

[project.scripts]
qr = "qrcode.console_scripts:main"

[tool.poetry]
packages = [{ include = "qrcode" }]
readme = ["README.rst", "CHANGES.rst"]

# There is no support for data files yet.
# https://github.com/python-poetry/poetry/issues/9519
#
# data_files = [
#     { destination = "share/man/man1", from = [ "doc/qr.1" ] },
# ]

[tool.poetry.group.dev.dependencies]
pytest = { version = "*" }
pytest-cov = { version = "*" }
tox = { version = "*" }
ruff = { version = "*" }
pypng = { version = "*" }
pillow = { version = ">=9.1.0" }
docutils = "^0.21.2"
zest-releaser = { extras = ["recommended"], version = "^9.2.0" }

[tool.zest-releaser]
less-zeros = "yes"
version-levels = 2
tag-format = "v{version}"
tag-message = "Version {version}"
tag-signing = "yes"
date-format = " %%-d %%B %%Y"
prereleaser.middle = [
  "qrcode.release.update_manpage"
]

[tool.coverage.run]
source = ["qrcode"]
parallel = true

[tool.coverage.report]
exclude_lines = [
  "@abc.abstractmethod",
  "@overload",
  "if (typing\\.)?TYPE_CHECKING:",
  "pragma: no cover",
  "raise NotImplementedError"
]
skip_covered = true

[tool.ruff]
target-version = "py39"
exclude = ["migrations"]
lint.select = ["ALL"]
lint.ignore = [
  # Safe to ignore
  "A001", # Variable is shadowing a Python builtin
  "A002", # Function argument is shadowing a Python builtin
  "ANN", # Missing Type Annotation
  "ANN401", # Dynamically typed expressions (typing.Any) are disallowed in `**kwargs`"
  "ARG001", # Unused function argument (request, ...)
  "ARG002", # Unused method argument (*args, **kwargs)
  "ARG005", # Unused lambda argument
  "D", # Missing or badly formatted docstrings
  "E501", # Line too long (>88)
  "EM101", # Exception must not use a string literal, assign to variable first
  "EM102", # Exception must not use an f-string literal, assign to variable first   "ERA001", # Found commented-out code
  "ERA001", # Found commented-out code
  "FBT", # Flake Boolean Trap (don't use arg=True in functions)
  "N999", # Invalid module name
  "PLR091", # Too many statements/branches/arguments
  "PLR2004", # Magic value used in comparison, consider replacing <int> with a constant variable
  "RUF012", # Mutable class attributes https://github.com/astral-sh/ruff/issues/5243
  "SIM105", # Use contextlib.suppress(ImportError) instead of try-except-pass
  "SLF001", # private-member-access
  "TRY003", # Avoid specifying long messages outside the exception class

  # Should be fixed later
  "C901", # Too complex
  "N802", # Function name should be lowercase
  "N803", # Argument name should be lowercase
  "N806", # Variable should be lowercase
  "PERF401", # Use `list.extend` to create a transformed list
  "S101", # Use of 'assert' detected

  # Required for `ruff format` to work correctly
  "COM812", # Checks for the absence of trailing commas
  "ISC001", # Checks for implicitly concatenated strings on a single line
]

[tool.ruff.lint.extend-per-file-ignores]
"qrcode/tests/*.py" = [
  "F401",  # Unused import
  "PLC0415", # Import not at top of a file
  "PT011", # pytest.raises is too broad
  "S101", # Use of 'assert' detected
  "S603", # `subprocess` call: check for execution of untrusted input
]
